+++ extensions/hello/src/App.jsx
@@ -1,11 +1,29 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const Wrapper = styled.h3`
-  margin: 8rem auto;
-  text-align: center;
-`;
+import React, { useState, useRef } from 'react';
+import { Loading } from '@kubed/components';
 
 export default function App() {
-  return <Wrapper>Hello hello</Wrapper>;
+  const [loading, setLoading] = useState(true);
+  const FRAME_URL = 'http://hello.www.ks.com';
+  const iframeRef = useRef();
+  const onIframeLoad = () => {
+    setLoading(false);
+  };
+
+  return (
+    <>
+      {loading && <Loading className="page-loading" />}
+      <iframe
+        ref={iframeRef}
+        src={FRAME_URL}
+        width="100%"
+        height="100%"
+        frameBorder="0"
+        style={{
+          height: 'calc(100vh - 68px)',
+          display: loading ? 'none' : 'block',
+        }}
+        onLoad={onIframeLoad}
+      />
+    </>
+  );
 }
