+++ extensions/hello/src/routes.js
@@ -1,0 +1,10 @@
+import App from './App';
+
+const routes = [
+  {
+    path: '/hello',
+    element: App,
+  },
+];
+
+export default routes;
